Online Student Scheduling: Banner XE Interface
- credit checking: maxHours changed to float (was int)

+++ JavaSource/org/unitime/timetable/onlinesectioning/custom/purdue/XEInterface.java
@@ -157,8 +157,8 @@ public class XEInterface {
 		public Boolean validStudent;
 		public String registrationException;
 
-		public Integer minHours;
-		public Integer maxHours;
+		public Float minHours;
+		public Float maxHours;
 	}
 	
 	public static class CourseReferenceNumber {
